[MS59.1] Add NetworkStatsDataMigrationUtils
Once the data usage part moves into the module, the OEM would
lose the ability to modify the code, thus we need a mechanism
to prevent data loss during migration. Hence, provide a
helper class to read old version of persistent network
statistics, the implementation is intended to be modified
by OEM partners to accommodate their custom changes.

This interface will be invoked by NetworkStatsService during
the boot up phase in the migration process which will be
introduced in follow-up changes.

Test: atest NetworkStatsDataMigrationUtilsTest
Bug: 204830222
Change-Id: I4a9c8a04761b530ce0b31869f0d7fc191b00438f

Former-commit-id: 44f61309c9daaa5f0ba7a974ecf79f31d3e95a09