Create framework-connectivity shared filegroup
Since data usage related code is going to move to mainline module,
hidden classes and methods can no longer be accessed. However,
some of the hidden classes in the platform are still needed, and
cannot be moved because of other callers.

Thus, create a shared srcs filegroup to allow these classes
build with the module to solve the dependencies.

Test: TH
Bug: 204830222
  (cherry-picked from ag/16656704)
Change-Id: I3813bc2177b2f0b4237056f64c57c63fa29c3081
Merged-In: I3813bc2177b2f0b4237056f64c57c63fa29c3081

Former-commit-id: 0580abcc9acfa5d5e0967d547263a10c3399d698