Animate dialog to app launches (1/2)
See b/214961273#comment1 and #comment2 for before/after videos as well
as technical explanations on how this animation works.

This CL also enables the generation of Java 8 default method bodies in
interfaces, in the animation library. This is necessary to avoid
unexpected behaviors when using @JvmDefault on an interface +
delegation, which is a common pattern when using the animation library.
See [1] and [2] for more info. Note that this is enabled only for the
animation library (and not SystemUI(Lib)) to avoid breaking code
depending on the (now deprecated) @JvmDefault method in SysUI.

[1] https://youtrack.jetbrains.com/issue/KT-34612
[2] https://blog.jetbrains.com/kotlin/2020/07/kotlin-1-4-m3-generating-default-methods-in-interfaces/

Bug: 214961273
Test: atest DialogLaunchAnimatorTest
Test: Manual
Change-Id: I27d53400a5dc852543a68f5b6f107d9eb08b06db

Former-commit-id: 6e77c58de86cb973030851ef435da965e429e5a2