Handle touch focus when clicking toolbar window transparent area
We use a fix size window to hold the toolbar no matter if we show it
horizontally or vertically. We should transder the touch focus to the
application behid the toolbar when clicking toolbar window
transparent area.

Bug: 190030331
Bug: 214158202
Test: manual. The toolbar can dismiss when clicking the transparent
area in this fix size window.

Merged-In: I3915e8165b4d371700cb5fb9e7eb23d46f39bd5c
Change-Id: I73bdec96303edab77acc55af3d7871d504614be7

Former-commit-id: b0040b03c22c2daa0c3a8cc10aac8a2fa9a73c24