Device state based auto-rotation setting: add fallback setting mechanism
This is required so that when the device goes into a state where the setting is ignored (e.g. half-folded), it falls back to the setting that is specified in the config, instead of simply the last state that was set.

This helps apps that read the auto rotation setting to make decision changes to get the correct value.

Bug: 203189552
Test: DeviceStateRotationLockSettingControllerTest
Change-Id: I3c6b006dc7ffdcfff62798558f09599f63a9f70a

Former-commit-id: d41d1a691a3c31a106b9d25a9981a80bb7cac062