Inline InputMethodMenuController.HardKeyboardListener
This is a follow up CL to our previous CL [1], which introduced

   InputMethodManagerService.HardKeyboardListener,

which was later moved to

   InputMethodMenuController.HardKeyboardListener

as part of refactoring [2].

This CL mechanically inlines it to InputMethodManagerService as it
allows us to achieve the same behavior with fewer lines of code.

This is mechanical refactoring. There should be no observable behavior
change.

 [1]: Ica768083f95c33dc1e494a28ba7d8b6eb989b0ef
      c08ab5b85676cc760447fa37e66f9c66f22bdae3 [formerly 01038492ff0317f0d3cff54d8a7ee36bb31ff175]
 [2]: Ief3c691029b14d848f6ad4a5ec8a8743760c7cda
      d81da158a192ab7721a59894c67fe0259294c0ef [formerly c88452c3ee33b8a517881bb2504f5e1c1126372f]

Bug: 192412909
Test: Manually verified by attaching/detaching a hardware keyboard.
Change-Id: I9cce75011a8564c9da8307c5555d87bcba4db2cb

Former-commit-id: 06c3998532734b58c0c32d1ae4d5786c764733d7