Rename ServiceRecord.isolatedProc to isolationHostProc.
This member is used to keep track of a process that we have started for
hosting a service. For normal services, we don't need to keep track,
because we can identify the process that should host the service by
matching its uid with the uid off the service. This however is not true
for isolated processes (which have their own uid), and now it's also not
true for supplemental processes (which also run in their own UID range).
To make clear that this will no longer just be used for isolated processes,
rename to isolationHostProc.

Bug: 215012578
Test: N/A
Change-Id: Ida90829ae196bf1161892e3da822444c913885ba

Former-commit-id: 174211109ccfa047d80d5df571de3c3579b95768