Add avatar picker that shows a predefined set of images.
This is the first in a series of CLs to add a new avatar picker.

This CL adds the basic avatar picker screen to show a set of
predefined images (that can be supplied as an array of drawables).

Resources are provided to adapt the screen to phones and tablets
of various sizes.

Future CLs will integrate the take photo and choose photo
functionality from the current dialog, and will also add a set
of default icons that can be chosen if no predefined images
have been provided.

Bug: 215134398
Test: manual check screen on various screen sizes
Change-Id: I6ce95549cc17dc5a9b67d9dc43c5b7d2bda37874

Former-commit-id: f98ae93fdf751e5b2326b1f78599a87e0be028a7