Removing RWC#getValidLaunchRootTaskInTaskDisplayArea
... and consolidate the flow with TDA#getOrCreateRootTask.

Updating few changes regarding to the launch windowing mode
calculation since the windowing mode is an important input
while determining the root task. It wasn't needed because
the candidate root task was reused in most cases.

Secondly, the root task was always created on default TDA
(regardless the preferable one is the secondary TDA) if no
launching activity was given. This CL also changes the behavior
to allow creating root tasks from the preferable TDA.

Bug: 184806710
Test: atest RootWindowContainerTests
Change-Id: I1e8055bb7bb2da8fe35bdfae1f85c610d5751434

Former-commit-id: bbcd34357beafefa2fc648094ae4a025e674d895