Ensure that "smart" actions are shown in Talkback.
The "Local Context Menu"
(https://support.google.com/accessibility/android/answer/6007066?hl=en-GB)
is missing smart actions. This CL adds them.
We need this to happen even more now that we are going to be rendering
the selection toolbar in a system UI and are experiencing accessibility
focus issues. See referenced bugs for more details.

Bug: 214122485
Test: m
Test: https://screenshot.googleplex.com/3SdNNBt5bWMz2CN.png
Change-Id: If5266c8b44461349877cf0029b7803e0bc8a9174

Former-commit-id: 6487bc38c174135e49424bf2d96ea3e77e77a585