Add required flags to registerReceiver calls in sysui
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
non-system broadcasts. This commit adds the RECEIVER_EXPORTED
flag to the receiver in DebugModeFilterProvider since this broadcast
is expected from the shell for debuggable builds, and the
RECEIVER_NOT_EXPORTED flag to the receiver in ControlsControllerImpl
as this broadcast is sent from the local app.

Bug: 161145287
Test: Build
Change-Id: I346b6ce4e1dd4e4d2f07953c284c5a1809982e35

Former-commit-id: 16ee1da74ecf4b402810665ef2d6f5934404fe50