Clean up foreground service manager controllers
Use the same controller for the UI and the dialog and remove the
RunningFgsController since it was very specific in what it provided.

Also have the dialog disable the stop button when apps are stopped while
showing. Also create exemption list for which apps can't be stopped or
shouldn't even be visible.

Test: Manual
Bug: 201579707

Change-Id: I1e528acaa01d539d89163d7241fd0aba9c19a82f

Former-commit-id: 9deea4a4357e17c6c3cfd4b4884abf13c7873b47