Add a GameSession API for restarting the associated game.
Since restarting a process must be done from the system server side,
we accomplish this by passing a IBinder through to the GameSession
allowing the GameSession to make IPCs to its hosting
GameServiceProviderInstanceImpl. As an additional layer of security,
the implementation takes a taskId and only restarts the given
game if the taskId matches an active game session.

Test: atest FrameworksMockingServicesTests and manual e2e testing
Bug: 210127167
Bug: 202414447
Bug: 202417255
CTS-Coverage-Bug: 206128693
Change-Id: I944da060323ac1779cd2c61d1e411e9868caed91

Former-commit-id: 614ba1384712818f74b256a8545d6804d3fccbb4