[MTE] show notification if using MTE bootloader switch.
ro.arm64.memtag.bootctl_supported can be set by OEMs that do not want to
enable MTE by default yet but want to offer users a preview that can be
enabled in the Developer Options. This developer option will reboot the
phone and enable MTE. In that case we want to show a notification to
make sure the user does not forget that it is enabled.

Bug: 206895651
Change-Id: Iaebc16d4b8d0da302f18709e23062f4a5bd58fe3

Former-commit-id: c3d0e6a8d265f1c131eefe0ac4c077b49c93598c