Enable Shortcut integration with AppSearch.
Static shortcuts and dynamic shortcuts will now be additionally
persisted in AppSearch in a background thread when they are published.

Shortcuts persisted in AppSearch have a TTL of 90 days and will be
removed if they are not republished (i.e. used by the user) within the
timeframe.

The number of shortcuts persisted in AppSearch are not bounded by the
limit defined in ShortcutManager#getMaxShortcutCountPerActivity.

Bug: 151359749
Test: atest CtsShortcutManagerTestCases
Change-Id: I6b1474bf60529e43172a0facae5561d2d89ede46

Former-commit-id: 523f3050065010afb4a0b7d409a9a062535f6b29