[RESTRICT AUTOMERGE] UserDataPreparer: reboot to recovery if preparing user storage fails
StorageManager.prepareUserStorage() can throw an exception if a
directory cannot be encrypted, for example due to already being
nonempty.  In this case, usage of the directory must not be allowed to
proceed.  UserDataPreparer currently handles this by deleting the user's
directories, but the error is still ultimately suppressed and starting
the user is still allowed to proceed.

The correct behavior in this case is to reboot into recovery to ask the
user to factory reset the device.  This is already what happens when
'init' fails to encrypt a directory with the system DE policy.  However,
this was overlooked for the user directories.  Start doing this.

Bug: 164488924
Bug: 224585613
Change-Id: Ib5e91d2510b25780d7a161b91b5cee2f6f7a2e54
(cherry picked from commit 8bbdf2988ac91268c0228929889137908642c6d7 [formerly 5256365e65882b81509ec2f6b9dfe2dcf0025254])
Merged-In: Ib5e91d2510b25780d7a161b91b5cee2f6f7a2e54

Former-commit-id: 69c3ce70c6dcabf57219d338af86e569ea672ef5