Change default LocationRequest quality
In order to maintain backwards compatibility with SystemApi versions of
LocationRequest, we chose to keep the default quality level as low.
However, this is causing problems when LocationRequests are created
through the use of legacy public APIs, leading to location requests that
do not receive locations as expected. Default the quality level to
balanced to address this issue.

Also removes old UnsupportedAppUsage annotations (past the 2 release
support limit now), and updates some variable names accordingly.

Bug: 215186921
Test: presubmits
Change-Id: I6e1fdb0b94438fd3dede75103cd70443e863a3f2

Former-commit-id: abb7ac8e60932c63662b36577f6722ab1316755f