Merge system and module-lib APIs from conscrypt and i18n
Previously, conscrypt and i18n were excluded from the list of modules
from which combined_apis retrieved system and module-lib API files as
they did not provide those files. They now provide those files so no
longer need to be treated specially.

The conscrypt and i18n (and art) are excluded from the
framework-updatable-stubs-module_libs_api as that is used with a
java_system_modules that already provides the conscrypt and i18n
and art APIs.

Bug: 216435117
Test: m all-modules-system-stubs
      m frameworks-base-api-system-current.txt
      m frameworks-base-api-module-lib-current.txt
Change-Id: I58502249c59b1faa65fb288ac11becd9496ff37e

Former-commit-id: 57533b4d8341023e3a6970434c33e9a19be9907b