New Pipeline: improve "old pipeline running" handling
Currently, we show a toast and log a message with a stacktrace when old
pipeline code is running while the new pipeline is enabled.

With this change, we show a toast *only on debug builds*, and we either
log a message *or throw a RuntimeException* depending on the state of a
new System UI flag.

Bug: 216092204
Test: manual
Change-Id: I6e4d27704e4806a3ea4dc0c20c4f727b1a41ac63

Former-commit-id: eb7d677acbd0f72262bd79ceb8cfb3927d52ebdf