"rounded_corner_content_padding" as a FW resource
This is a follow up CL to my previous CL [1], which introduced

  input_method_nav_content_padding

as a copy of SysUI's

  nav_content_padding.

What we found since then is that instead of trying to keep two
resources in sync it'd be more easier if we have

  nav_content_padding

as an internal Framework resource.

With this CL, SysUI's one becomes a synonym of the FW's one.

Note that existing device-specific overlays for SysUI resources
continue to work hence there should be no observable behavior change
in this CL.

In subsequent CLs, we are going to update each device-specific
resource overlay to override FW's one.

 [1]: I3e7e1f83554444131e2765dc159617bb9e2337c7
      25d9e4edafa41809f96d4de0ad1c43559c44afbd [formerly ff7b453ca8e23b4ef75ba2c3f5becaf511cb07d3]

Bug: 215554582
Test: Manually verified that the nav button locations are the same
Change-Id: I0c3717c0bdfc928167c6eba06247b37caabf33b9

Former-commit-id: 24d6b891f572f7be419d3acc9793ca68d062cf35