Updates OverrideRequestController to allow for one override request at max
To simplify orchestration around handling multiple override
requests, we are reducing the OverrideRequestController to
only having one override request at a time. If a new request
is added, the previous request will be cancelled and removed.

This change also means that we can remove the SUSPENDED status
of a request and the accompanying callbacks

Bug: 216495664, Bug: 207686851
Test: OverrideRequestControllerTest
Change-Id: I787f2ce8567812a633c0afdd5b3f684da4508249

Former-commit-id: cda48cdca5142546e5f8c275b10b80fd4e7aaa8a