Remove checks for FDE in 'adb backup'
The deviceIsEncrypted() methods in BackupRestoreConfirmation.java and
UserBackupManagerService.java only return true if the device is using
FDE (Full Disk Encryption), for which support has been removed in favor
of FBE (File Based Encryption).  Therefore, the logic to require a
backup password no longer applies to any device.

Remove this logic to simplify the code.

It is possible that this was actually a bug, and this logic should have
applied to FBE devices too.  But given that the code has worked this way
for years, and there isn't necessarily a logical connection between
whether the device is encrypted and whether a backup *must* be
encrypted, I decided not to change the current behavior.

Bug: 208476087
Test: 'adb backup' and 'adb restore' still work.
Change-Id: Idc72d2a4c3e8bfa10a32cdc57884159b37635e81

Former-commit-id: d5b040ed643948dd892f4ad8e2d0f3cfa8d4613d