IME Support for SurfaceControlViewHost
Since SurfaceControlViewHost embedded hierarchies can already receive
focus via grantEmbeddedWindowFocus, and ViewRoot can communicate
directly with IME service, most of the wiring is in place for IME
to work with SurfaceControlViewHost. It's mostly a few validation
checks around focus preventing us from doing this. To solve this we
make ImeInputTarget a "InputTarget" reference rather than a window
state reference. We extend InputTarget to allow for abstraction of
these few cases. At the moment we force embedded views to use
the remote insets target from SysUI, and they are not allowed
to control them themselves.

Bug: 213603716
Test: SurfaceControlViewHostTests
Change-Id: I7eec5bdf6798ef8cc5e6958987ca7b73b6881079

Former-commit-id: 205da89cafb811db24cdc8d73e49f27dfd2656c8