Fixing duplicate status bars in split shade
The issue was updating view state more often than before - in updateQsExpansion (introduced in ag/16551189).
As reaction on StatusBar state change - going from Keyguard to Unlocked, first NPVC hides KeyguardStatusBarView but then, through chain of calls it calls updateQsExpansion and again ask for update KeyguardStatusBarView state. The issue is we're still in StatusBar listener call chain and the same listener in KeyguardStatusBarViewController hasn't been called yet - so it still thinks we're in Keyguard states and makes view visible again.
This issue can be solved in a few different ways, but this refreshing view state was introduced just a few days ago and for functionality that is disabled now so seems like the safest fix.

Fixes: 216104488
Test: manual
Change-Id: I7dfdf91f397bf5077c639248de702596014fd8e6

Former-commit-id: 5b4247d0ccf82fb205c730d62ed2fbc5c70b1fe1