Add 'miniresolver' UI for cross-profile single app resolution.
When the current profile only has web browsers and the other profile has
a single non-browser app to handle the given intent, simplify the
resolver UI to present the user with a yes/no question about how to
handle.

Canonical example of this is "clicking a youtube link in your work
profile".

Bug: 175433480
Test: atest ResolverActivityTest
Change-Id: I213313ca87a8d1db5e586fe074f953060a91fc19

Former-commit-id: 81e474d7a73516faa1e45a6279020471aa453232