Include all apps when getting Recents
The current RecentTasks fetches the recents app (based on
config_recentsComponentName). However, it does so for user 0. In some
cases, the recents app might not be installed in user 0 (even though it
is installed in other users). We need to fetch it even in that case, so
that recents can work on those other users. (Otherwise, the attempting
app will not only not do recents, it'll probably crash.)

Bug: 216636343
Test: manual confirmation on a headless user 0 device where launcher is
uninstalled on user 0

Change-Id: Ib2b536e997d58f8fb34c75e92a1cdd843b4b04fa

Former-commit-id: 992f8b5b0fb563c79fe3db86b930022b2c3810d6