Notifies magnification change for the transitioning target mode
When the magnification settings mode or config mode transitions,
MagnificationController would
1. Disables the current magnification mode. Then
2. Starts to enable the target magnification mode.

To prevent the magnification change callback send twice during the
mode transition, MagnificationController should only notify the
change of the transitioning target mode and ignore the notifcation
change for the disabling magnification mode.

And when WindowMagnification resets or disables by
WindowMagnificationController#deleteWindowMagnification(),
it should notify source bounds empty when magnification is deleted.

And when FullscreenMagnification resets and the animator is animating,
FullScreenMagnificationController should notify magnification
change.

Bug: 209701404
Bug: 210636326
Test: atest AccessibilityMagnificationTest,
    atest WindowMagnificationControllerTest,
    atest com.android.server.accessibility.magnification
Change-Id: Ief718614d80177a4a076eddc708a7ecfc4dfb73f

Former-commit-id: d560a73ca9e38ffbf8a7115d1f09814f0c070b2f