Animate notification shade during unfold
Adds some constant translation to elements of the notification shade during fold/unfold. This is only applied when the device is in natural rotation.

KeyguardUnfoldTransition was doing the same: the logic has been extracted in UnfoldConstantTranslateAnimator, that is now used by both classes.
KeyguardUnfoldTransition's tests related to the animator have been moved to UnfoldConstantTranslateAnimatorTest.kt.

+ Minor formatting issues fixed (using ktfmt)

Bug: 201411030
Test: atest UnfoldConstantTranslateAnimatorTest KeyguardUnfoldTransitionTest + Tested manually on lockscreen and notification shade
Change-Id: Ic8d4cf5d7b2a9f8fc366842623c87c15cf79766f

Former-commit-id: 4b40d32ae4556fc1b9103a3c405f1b96e50fe993