Added API to set active assistant UIDs
Added API to manage the currently active assistant. When an
assistant becomes active the API will set the assistant's
voice interaction service and its corresponding hotword detection
service to active. When the assistant becomes inactive the API will
be called to reset the assistants UIDs. For the latter the API will
be called with an empty list of UIDs.

After bootup, user switch, and audio server crash reset, the list of
active assistants will be set to an empty list (.i.e no assistant
UID's are currently active).

In the absense of active assistant list, the default behavior will
be that all assistants can listen to hotword as long it does not
conflict with privacy policy. When an assistant state changes
to top (UI present) that assistant will be able to listen to
audio via voice recognition source or hotword.

Bug: 189312611
Test: m -j, and run assistant
Test: atest CtsVoiceInteractionTestCases
Change-Id: I0bfe4c167f69a26293ae2f9f9d0898033f86db03

Former-commit-id: 3ef3f58c9862669779558de47c3d8b04e20fd7ae