Animate power menu dialog from power button
This CL animates the power menu dialog from the power menu button
instead of from the bottom. The animation has the same spec as the
device default dialog animation (see http://ag/16769344), except for the
direction of the translation.

Note that this does not support devices that don't have the power button
on the right. I will add support for that in a follow-up CL.

See b/213872558#comment3 for a video.

Bug: 213872558
Test: Manual
Change-Id: I7d5f7d9f2ed2375814cb827275728ef9890a50ae

Former-commit-id: 0b9512f6e88bf752d69f89e7d9e9195835d36c85