Fix the empty selection issue when onCreateActionMode returns false
One of the optimizations ag/12911059 did was calling
SelectionModifierCursorController.show() before
startActionModeInternal(). The rationale was that if we start the action
mode first, SelectionModifierCursorController.show() would end up
invalidating the action mode twice unnecessarily, once for each handle.

However, with this optimization, we are calling
SelectionModifierCursorController.show() even when onCreateActionMode
returns false.

Reverted this particular optimization to fix the issue.

Added a test which was failing without this fix but passing with it.

Fixes: 199380016
Fixes: 214341747
Test: atest TextViewActivityTest

Change-Id: I793f76a23978cbbbbde2d16e8a522615174bcdd5

Former-commit-id: 11bd64482249c9257ee22390de15f69f9c834247