Clarify which InputMethodManagerInternal method takes lock internally
InputMethodManagerInternal is the callback where we need to pay extra
attention to potential dead-locks.

To make code reading easier, this CL makes it clear about which method
needs to acquire ImfLock and which method is always thread-safe.

Basically this is a mechanical code refactoring.  There should be no
observable behavior change.

Bug: 192412909
Test: presubmit
Change-Id: I98958e6c149582b0a24997c9ca814e355f8b2a2b

Former-commit-id: 0fee8d1035383e8cc046d8e3f2f458fbf3c30d52