DisplayManager: Make sure RampAnimator sets property in linear space.
This fixes the case where RampAnimator is used the first time and
immediately jumps to the target brightness. In that case the property
was set to a value that has not been converted back into linear
space, causing an incorrect initial brightness value.

To avoid these bugs, consolidate setting of the property into one
function.

Bug: 217289549
Test: Cannot reproduce bug.
Change-Id: I2ccdaa51ddba062fc5de1209470b90eb429280db

Former-commit-id: 246d553f51499658d45c2eedaf79a51f7b540dba