Drop input for toast and child surfaces
Toasts that do not have the trustedOverlay flag should not receive input.
These windows should not have any children, so force this hierarchy of
windows to drop all input by setting a flag on the toast window state
which will apply the DROP_INPUT flag on all windows with an input
channel. This is to prevent malicious apps from parenting surfaces with
input channels to the toast window.

Test: show toast and check if input feature flag DROP_INPUT id set via dumpsys
Bug: b/197296414

Change-Id: I316b76b685ca5030fd8aa91283555efcce4d6994
Merged-In: I316b76b685ca5030fd8aa91283555efcce4d6994

Former-commit-id: 55c1473bf2fedaacf7bb8ac068e6f9b1a625b5e0