New Pipeline: plumb isChangingPosition
ExpandableNotificationRow has an isChangingPosition flag that's used to
inform the remote input code of when a view is being detached only
briefly as the shade is rearranged, to give it a chance to save and
restore state that would otherwise be discarded.

The new pipeline doesn't currently set and clear this flag, which broke
lifetime extension for notifications with a remote input history (one or
more replies sent), but also kept input focus from being preserved when
a notification is moved.

This change sets and clears the flag in the relevant NodeControllers. It
may or may not render ag/16749051 redundant.

Bug: 211161595
Test: manual, after locally reverting ag/16749051
Change-Id: Ie808a854f6a0bc523dffcc2787818e5765d507ec

Former-commit-id: 4d4e94ca9450701907ed41d0c953c10b8b2fb337