Fix potential for dead binders when calling callbacks
This patch ensures that all callbacks out of the VcnManagementService
are try/catch guarded to ensure that callbacks are not called and throw
an exception.

This is triggered when telephony dies, which results in the
subscriptions changing, and attempting to notify the policy listeners
triggers a DeadObjectException.

Bug: 217537578
Test: atest FrameworksVcnTests
Original-Change: https://android-review.googlesource.com/1969041
Change-Id: Ibfe82fb187b7989164e22919a72b8c36d2655902
Merged-In: Ibfe82fb187b7989164e22919a72b8c36d2655902

Former-commit-id: 4975dd39e835d4bec80c5bca0e2cf1f916951bf4