Fix tile request dialog dismiss
When the dialog is dismissed without being cancelled (for example, by
going home or locking the phone), make sure to also send a dismissed
message. As dismissed is always called after cancelled or click in a
button, make sure that this message is only sent once.

Also, check for RemoteException when communicating with the Binder
interface.

Test: atest TileServiceRequestControllerTest
Test: manual with test app
Test: CtsVerifier
Fixes: 217887724
Change-Id: I286c24ae743aede158214880679abd07c3db3ff6

Former-commit-id: 3cf1b970d40da9fdc9f425f03d79c570107f80c3