Adding a timeout for active-unlock
Adding a timer for active-unlock. This timer should interact smoothly
with the timer for extend unlock. That means that if there is a timer
for extend unlock that expires while the active unlock timer is running,
it should set the phone back into a trustable state. If there is an
extend unlock timer when the active-unlock timer expires it should
prevent active-unlocking but allow extend unlock to continue.

Test: manual
Bug: 215391978
:

Change-Id: Ia0ccf7ecc9dacf9f4a7a913c7fce8532e6a1ece1

Former-commit-id: 2410cfc05e226d1c1f6f9364764d03c6f149b5b8