Put security footers in new footer
This CL puts the security footer and fgs disclaimer in the new footer
(based on flag) instead of QSPanel.

The FGS disclaimer has two modes. A text mode (when security footer is
not visible) or just a number with maybe an update dot (when security
footer is visible).

Also, use DialogLaunchAnimator for QSSecurityFooter

Test: manual
Test: atest SystemUITests
Fixes: 217908230
Change-Id: Id9289826de3256ea6fdfea3ca6f8bf1976364e61

Former-commit-id: d68c626be06be2770ff52dc0a2473f1d9b90f74f