Improve compaction by skipping bad VMAs instead of fully bailing out
Previously when a VMA failed compaction for any reason the system would
stop trying to compact all the rest of the VMAs and bail out. However,
there are multiple reasons that a VMA can fail due to not being
reclaimable with -EINVAL, in such instances we will just skip the VMA
and keep going with the rest of the VMAs and any other error will still
cause compaction to bail out as it would likely be irrecoverable.

Test: Manual. Verified that once a VMA errors with -EINVAL, it continues
with the rest of the VMAs and other errors bail out.
Bug: 205658049

Change-Id: Ifc190e371c4dce0eaa6dbab104aa0b666e06027d

Former-commit-id: cbb6b2be373dcd44a3705937745dd3f66196c5a3