Introduce config_imeDrawsImeNavBar
This is a follow up CL to my previous CL [1], which used the following
conditions to determine when to let the IME render the back and IME
switcher buttons.

 * InputMethodService#canImeRenderGesturalNavButtons(), and
 * com.android.internal.R.integer.config_navBarInteractionMode == 2

In production, there may be cases when we want to use the previous
mode even if config_navBarInteractionMode remains to be 2.  As a
preparation to support such scenarios, this CL introduces

  config_navBarInteractionMode

as an overlayable resource.  In subsequent CLs, we will start actually
using it.

Anyway there should be no observable behavior change in this CL.

 [1]: I3e7e1f83554444131e2765dc159617bb9e2337c7
      25d9e4edafa41809f96d4de0ad1c43559c44afbd [formerly ff7b453ca8e23b4ef75ba2c3f5becaf511cb07d3]

Bug: 216118048
Test: Manually verified as follows
 1. Build aosp_coral-userdebug and flash it
 2. adb shell cmd overlay lookup android android:bool/config_imeDrawsImeNavBar
     => "false"
 3. Enable gestural navigation
 4. adb shell cmd overlay lookup android android:bool/config_imeDrawsImeNavBar
     => "true"
Change-Id: I0651a99b6007a84be63b85b579a85af0f24bb6ec

Former-commit-id: b89cbc6f05af0bb1637e6992835b6fc00ed8be8c