[10/n] Letterbox Education: don't show if taskbar education is showing.
Introduce a LAUNCHER_TASKBAR_EDUCATION_SHOWING setting for taskbar to
indicate that the education is about to be shown or is currently
showing.

After the taskbar education is dismissed, the letterbox education will
only be shown for a different task or for the same task once its compat
state changes and the letterbox education is eligible to be shown.

Bug: 214590804
Test: N/A
Change-Id: Ia8512e5f188fe421adf1de2fc8a0422b9cc79c76

Former-commit-id: d0df97d939c6ef81bff7d841cea4a38cf259a369