[Media TTT] Add app icon and name to the receiver SystemApi.
We typically get the app icon and name using
`context.packageManager.getApplicationIcon(mediaRoute2Info.packageName)`.
However, this may not work on the receiver device if the receiver
doesn't have the app installed.

The API will allow the caller to provide the app icon and name directly
so that we can ensure the chip on the receiver device is displaying the
necessary app information.

Bug: 216318437
Bug: 217418566
Test: atest MediaTttChipControllerReceiverTest
Test: CTS tests updated in different CL in same topic
Change-Id: Ib5150355a08ae389b218500b235b4bd9c686f919

Former-commit-id: 27378ee36853fe99545b29d9171eb354fc20fd2c