Introduce ImePlatformCompatUtils
To easier maintain more platform compat behavior changes in IMF side,
with wrapping IPlatformCompat checkings into a new utility class that
used by IMMS to check if the given state is valid to enable/disable
IME behaviors on the platform side.

This is a mechanical refactoring and should not have any behavior
changes.

Bug: 16825711
Test: Build
Test: atest CtsInputMethodTestCases
Change-Id: Icc19e52567de58cd72f2abba1b0aa6dd176995fd

Former-commit-id: a97ad615016db2ee19b8918d8f37baf6a65a310a