Correct package installation behaviour in GameManagerService
Previously when package is changed, GameManagerService always updates
the configuration, however, this is only necessary when package is
installed and removed. Hence correct the behaviour here to avoid
unnecessary. In the meantime, when a package is removed, also remove the
game mode set on the package because carrying the game mode for
uninstalled package is also unnecessary.

Bug: b/199920928
Test: atest GameManagerServiceTests
Change-Id: Idfd5c78776c182b37ef4c969991b1a4d88e65390

Former-commit-id: 7043659eb4a8ef6c65c850a7173ce73d51628190