A overloaded API for downloadSubscription with portIndex for LPA
Most of cases, download opperation is port agnostic we don't need
to specify the portIndex. However if this is triggered from carier
apps and apps specify download followed with esim activation, then
platform need to resolve a portIndex(maybe through UX) and need to
pass this portIndex to LPA app. thus we need to add an overloaded
version of download API for LPA.

Bug: 214055002
Test: Build & atest EuiccServiceTest
Change-Id: If127224ca7ca696c8eb512dc80c27f54e4f25b38

Former-commit-id: 3c22236cdad635097a88e6c958188fec325ab43b