Fixing not responsive tiles in split shade
Some tiles (beyond QQS tiles) were not responsive because when QSPanelController is about to start listening, it also checks if mQsExpanded is set to true. mQsExpanded is passed from NotificationPanelViewController and it wasn't set to true because function setting it is never called in split shade.
This change is hacky but probably safer way to solve it by changing condition for listening on changes.

Bug: 218268829
Test: open split shade and tap night mode tile a few times - the tile must be positioned beyond first 6 tiles
Change-Id: Id6ad9f199964dd416a52c169a3e30713389479cc

Former-commit-id: d80184457b33efcf427a426efb31293f7428e009