Reevaluate game service when provider changes
The goal is to have the system server be responsive
to changes in the game service provider package
and reevaluate the GameService and GameSessionService
accordingly. So, for example, if the GameService
component is disabled and then becomes enabled,
the system server should detect that and bind
to the newly enabled GameService. This is a
likely real-world scenario if a game service
provider uses a server-side flag to enable
or disable the GameService component after
install time.

This change also fixes an issue where
GameService#disconnect() would never be called
because the service connector was always unbound
before the posted disconnect() call could complete.

Bug: 217215722
Test: atest CtsGameServiceTestCases GameServiceControllerTest GameServiceProviderSelectorImplTest
Change-Id: I5d0a603b98f341b956ea6abe82930fa13f4f7d0e
Merged-In: I5d0a603b98f341b956ea6abe82930fa13f4f7d0e

Former-commit-id: 0f5e540c2d551db59492c38930f562a587ca2754