Add tool for injecting tracing code into a method.
This tool rewrites the bytecode in the designated JAR files to produce
tracing calls on enter and exit, while making sure to close the tracing
span even on exceptions being thrown.

The idea is mostly to reduce the amount of noise within methods when
just trying to add some tracing.

Test: atest --host TraceInjectionTests
Change-Id: If6acb72f34cbb83d9b041a62ee3d8c2abf74b69e

Former-commit-id: 4fd5fae75fbfd44ddf6f978a9bd076372e0f0f6f