Skip duplicated ANR before enqueuing
Although ProcessErrorStateRecord#appNotResponding has similar condition
to check isNotResponding, it may not cover all cases which depends on
the timing to set the flag, especially if there are multiple ANR records
for the same process.

It also reduces the race while an ANR record is dequeued but the process
may be killed or restarted before dumping stacktrace.

Bug: 218840177
Test: atest AnrHelperTest#testSkipDuplicatedAnr
Change-Id: I10dd6678c2d288b6526cf97bf82e24f052fd89f5

Former-commit-id: 851813ad16e7da392c1ffb3a4d0555d1cb686b13