Support ANRs from windows that are not tracked by WM
Previously, ANRs for gesture monitors were reported using the pid of the
owner, and ANRs for all windows were reported using its input channel,
which was tracked by WM.

Now, there can be input windows that are not tracked by WM. We unify the
ANR reporting pipeline so that we first try to report an ANR using the
window's input channel. If the ANR reporting fails because the input
channel was not tracked by WM, we fall back on reporting ANR via the
pid of the window owner.

Bug: 210978621
Test: manual: Create a gesture monitor that sleeps while processing
input events, ensure there is an ANR.

Change-Id: If368ccf4a5cc676f64f3ed9de4f765b0abe63284
Merged-In: If368ccf4a5cc676f64f3ed9de4f765b0abe63284

Former-commit-id: 280b13eacf1364610b24d1d1a8a5d899dde9595b