Tweak screenshot animation for display resolution changed
The screenshot is taken in old resolution, so if the logical
display changed size, the screenshot should resize with it to
fill current screen size.

Also remove logic added in commit 338374 because now there is
letterbox configuration to show app in its requested orientation
on a non-rotatable device, which won't use this animation.

Bug: 217279239
Test: Loop commands:
      adb shell wm size 800x1600
      adb shell wm size reset
      There should be no obvious flickering.
Change-Id: Ifb4ee93aa0d101f68159a9d03806589cfea79080

Former-commit-id: d98c3d0c76b7c560e3db78b05a25d69450663538