startProcessLocked: Inherit debuggable flag for SDK sandbox
If the process is an SDK sandbox, we use the FLAG_DEBUGGABLE value of the associated app.
This allows for attaching debugger to a sandbox (and their SDKs) when
the app's manifest specifies android:debuggable is true. To achieve
this, we add an isSdkSandbox property to the ProcessRecord which is read
when the flags are being set.

Bug: 210499058
Test: manual (1. load app with android:debuggable, 2. load SDK, 3.
attach debugger in IntelliJ/Android Studio)

Change-Id: Id0c6a5914aaabd6db28b562fe17d24396b462779

Former-commit-id: 6e130b154c74a7a967480b0c650a75111bc9b871