Fix app hibernation spammy logging
With the force-stop check *always* checking for app hibernation state,
we now somewhat expect calls to AppHibernationManager before user states
are ready. We remove the current logging around isHibernatingForUser but
leave the logging in the other methods as we don't expect those APIs to
be called before the device is ready.

Also switch to using TextUtils.formatSimple for performance.

Bug: 219835296
Bug: 192311229
Test: atest AppHibernationServiceTest
Test: bug repro. No more spammy logs
Change-Id: I3dc94e3c2c63e0dcbb65ca551a2f3a27951d6a6a

Former-commit-id: faad4a7d37312e3d85cd26ac9ee75c8ef9fef35b