DO NOT MERGE Only clamp off brightness from DOZE_INITIALIZED
When the display times out, the power manager
briefly decreases the brightness of the display
before transitioning to AOD, so we don't want
to immediately increase the brightness to the AOD
brightness. However, for all other state changes
we don't need to take into account the screen
off brightness clamping.

Test: atest DozeScreenBrightnessTest
Fixes: 197060508
Change-Id: I8d25096f45cdb18717cc3d110bf5949f202c3982

Former-commit-id: da3eabb2060fed4ea0491cf63e4a39ef299528d6