Lint for @hide tags in comments
Errorprone will warn for normal javadoc tags in comments, but doesn't
know about @hide and so ignores it. This can lead to accidental exposure
of APIs when folks mistakenly use a comment block instead of a javadoc
block.

Also, adds some static dependencies to enable actually running the tests
on the host.

Fixes: 181140148
Test: atest --host error_prone_android_framework_test:com.google.errorprone.bugpatterns.android.HideInCommentsCheckerTest
Change-Id: I46a59075f93468fc0e57264bca0d86994ebb3d86

Former-commit-id: f320353b3627c34e26eaca3811c580ecaa62dedb