Keep showing wallpaper when starting an app immediatly after unlock.
Wallpaper is controlled by NotificationShade while the lock screen is
shown, otherwise by the launcher. We used to switch it when unlock
animation finishes, but it caused an issue when a user starts anotehr
app transition (e.g. launch an app by tapping an icon on a home screen)
before the unlock animation finishes.

We give higher priority on the launcher to NotificationShade, so that
this switch happens earlier.

Bug: 215685199
Test: Manual. Unlock the device and taps an app icon to launch it.
Change-Id: Iddf58c3a3966e285870711f39b73eee7dab804e0

Former-commit-id: f0a9429cc250f30f20375c9a6ebe9fbe0cb0082f