Fade in the UDFPS icon from HOME => AOD
Animate the fade in on the first signal from onDozeAmountChanged
when the unlockedScreenOffAnimation is running.

Remove the callbacks from the UnlockedScreenOffAnimationController
progress which starts the animation much sooner than when the AOD
UI starts to come in. It was also causing a pulsing bug where
the UDFPS bouncer would be reset and be out of sync with the scrims.

Test: atest SystemUITest
Test: manually transition between the home screen
and the AOD screen with udfps enrolled
Test: manually transition between the lock screen
and AOD with udfps enrolled
Test: Tap on a pulsing notification in AOD when
udfps is enrolled
Fixes: 220219675

Change-Id: Idfe6144d1d46443db52583b165a836c69f17b259

Former-commit-id: 99d2e63d4efc80a3566ac43dee57373141d814ce