Add a condition of ime should be visible on tracking typing focus.
We add a mIsImeWindowVisibleArray in WindowMagnificationManager to
indicate whether there is a ime in the displayId Display or not.

The element in mIsImeWindowVisibleArray would keep the ime's visibility
attribute which can be changed by onImeWindowVisibilityChanged callback.

Then, we add this element which should be true to the decision of
triggering tracking typing focus functionality.

Bug: 215813890
Test: Manually, I use log print to trace whether the mIsImeWindowVisibleArray
is set during AccessibilityController#setMagnificationCallbacks.
      atest WindowMagnificationManagerTest
      atest FullScreenMagnificationControllerTest
      atest MagnificationControllerTest

Change-Id: I14f172255e79d04428db86769ed49d679f2790f5

Former-commit-id: dd4bdc29dc390f9949b2cc5f2f43e68ce74bfa44