[Fold to AOD] Fix large clock blinking when folding a device
Fixes blinking when large clock becomes a small clock
when folding a foldable device: when there are notifications
and split shade is not active we should show small clock.
Clock size (large/small) was updated only in
NotificationPanelViewController#OnLayoutChangeListener
when folding or unfolding. This is called only after
the next layout pass, so sometimes we could see smaller
clock when folding a device.
Added calling clock update each time when split shade
mode activated/deactived.

We don't need to worry about the new layout for calculating
coordinates for updating clocks because it is the same
in both cases (mStatusArea.getTop is not changing and it
equals to the height of the small clock, the actual
movement is controlled by the translation).

Bug: 202844967
Test: fold/unfold multiple times when AOD enabled
Test: atest com.android.systemui.statusbar.phone.NotificationPanelViewControllerTest
Change-Id: I73ee110b6cbd05cfb4a01af10f0d5356ff0c16a2

Former-commit-id: c6824a4f5de484b6097e0f63ef6476ddf7bad023