New Pipeline: log sections in NodeSpecBuilder
We're seeing the "Silent" section header hang around after all
notifications in that section are dismissed, so log a bit about the
sections in NodeSpecBuilder.buildNodeSpec.

Logging is only enabled if Compile.IS_DEBUG, and is at LogLevel.DEBUG,
so this shouldn't slow down user builds or clutter logcat. (If you'd
*like* to see these messages, run `adb shell settings put global
systemui/tag/NodeSpecBuilder v`.)

Bug: 216150338
Test: manual
Change-Id: I8729377d32669980afea8895a3ec85bbd00d0b87

Former-commit-id: 65aa94aa08f0bb44aa5625f4c2e6a10c82679343