Disallow PAP authentication when MPPE is requested
MPPE cannot work if PAP is used as authentication, so it is not useful
to allow PAP authentication when MPPE is enforced: establishing the
tunnel would fail anyway with "MPPE required, but MS-CHAP[v2] auth not
performed".
Also users enforcing MPPE may assume that this means PAP will not be
used for authentication, so without this change MPPE enforcement gives a
false sense of security, as PAP uses plain-text credentials.

Bug: 201660636
Test: atest VpnTest
Change-Id: Ie318d45fe44294e97cf38da7f1834cf014cb4417

Former-commit-id: 997a4a39268b4f3af7ccc388269b5eb1972d3624