Generate user consent dialog for device log access
This change is to implement dialog instead of generating notification when device log access
is requested.

The LogcatManagerService checks if logd access is coming
from background or foreground.

If the requester is foreground, the LogcatManagerService
generates a dialog to obtain user consent. If the user grants access,
the LogcatManagerService approves full device log access. If the user
does not grant access, the LogcatManagerService denies full device log access.

If the requester is background, the LogcatManagerService denies full device log access.

Consent dialog times out if there is no user action for 5 minutes and
the LogcatManagerService denies full device log access.

This codebase is the continuity of the implementation ag/16618162 to
obtain user consent and improvement of ag/16673548

Bug: 193708292
Test: CtsVerifier
Ignore-AOSP-First: pending fix for logcat privacy issue

Change-Id: I708ca651a655fc5725dfac035ac27e7fd3ed27f6

Former-commit-id: 8e094fb3c712e8540b2614d399c4cca378d4a045