Set touchable region for recents input consumer with shell transitions
It seems like when shell transitions is enabled, the recents input
consumer's touchable region was not set up properly. It could have been
relying on the previous behavior of an input window being touch modal by
default. Now that we have removed touch modal from input, ensrue that
the touchable region is correctly set for the recents input consumer
when shell transitions is enabled.

Bug: 216806304
Test: presubmit
Change-Id: I93c07c992b0b4e9b61c694a18fae99da73deadef

Former-commit-id: 1a5cea7609f4e488b95860c152e08d63e7825f4e