Remove double-locking from PowerManagerService.onBootPhase
The PowerManagerService.systemReady have some code that needs to be
executed with the mLock held, and some that requirest it not to be held.
One this is the BatterySaverPolicy.systemReady, that uses the same lock
object as the PowerManagerService and explicitly required it not to be
locked at that point.

Moving the synchronized block inside the PHASE_BOOT_COMPLETED phase,
since the incrementBootCount also don't require the lock as it acquires
one as well.

Fix: 222098360
Test: manual
Change-Id: I1a25de3c7ba4be6d88d2654d42cdd496387391c2

Former-commit-id: 8c6a02a4e0cc20a6f1dc4314b15a7e7b1cea1ae7