Replace MainWindowSizeChange transaction with applyWithNextDraw
It's legacy code from the days of deferred transactions but still has
callers because it's convenient in it's own way. It has some details
relating to relayout that could get in the way of our plan to remove it
and so we refactor the primitive to make use of the applyWithNextDraw
primitive to make it easier to port everything to the async system.

Bug: 161810301
Bug: 175861051
Bug: 175861127
Bug: 200285149
Bug: 214315611
Change-Id: I369ff2fc8b4c730c4d31d9c1b89970bd295c32a3

Former-commit-id: 20620bc01b14b070700eec6b530c34496e0ae3a8