Limit launching placeholder to top TaskFragment
Currently, we launch placeholder Activity if Activity is not split
and matches placeholder rule. However, there's a scenario that
the Activity which matches placeholder rule is occulded by a fullscreen
Activity, but SplitController still launches placeholder on top of the
fullscreen Activity.

This CL verifies if the Activity located TaskFragmentContainer is on top
to launch placeholder. It also makes #getTopActiveContainer consider
Activity from other process by verifying #getRunningActivityCount.

Test: manual - Repro steps in bugs
Bug: 214175523

Change-Id: Id287d8c5235918b78ccd43bfddf9bb17ee0ff2e8

Former-commit-id: 3d419be76f1c17fd8adc9d8ac7d7e63af461152b